Fix some gcc warnings
- remove "sys/" from includes that are generic and should not be used with "sys/" prefix
- add some explicit casts for sprintf calls

+++ httpcomm.cpp
@@ -127,7 +127,7 @@ void HttpComm::requestThread(ChildThreadWrapper &aThread)
         "\r\n"
         "%s",
         contentType.c_str(),
-        requestBody.length(),
+        (long)requestBody.length(),
         extraHeaders.c_str(),
         requestBody.c_str()
       );
